Split agent runner output helpers
