+++ scripts/agent-runner-dry-run.mjs
@@ -1,10 +1,9 @@
 import {
   loadEvaluatedProject,
   parseArgs,
-  printHumanSummary,
   projectConfigFromArgs,
-  projectSummaryJson,
 } from "./lib/agent-project-queue.mjs"
+import { printHumanSummary, projectSummaryJson } from "./lib/agent-runner-output.mjs"
 
 const args = parseArgs(process.argv.slice(2))
 const project = loadEvaluatedProject(projectConfigFromArgs(args))
